Scge ml

+++ modules/local/make_circos_plot.nf
@@ -11,13 +11,17 @@ process  MAKE_CIRCOS_PLOT {
 
     script:
     """
-    cp ${projectDir}/bin/circos.conf . && \\
-    sed -i "s|__FILE_PLACEHOLDER__|$circos_input|" circos.conf && \\
-    /circos-0.69-9/bin/circos -conf circos.conf
+    if [ -s "$circos_input" ]; then
+        cp ${projectDir}/bin/circos.conf . && \\
+        sed -i "s|__FILE_PLACEHOLDER__|$circos_input|" circos.conf && \\
+        /circos-0.69-9/bin/circos -conf circos.conf
+    else
+        touch "${meta.id}_circos.png"
+    fi
 
     cat <<-END_VERSIONS > versions.yml
-    "${task.process}":
-        circos: \$(echo \$(circos -v 2>&1) | sed 's/circos.*v //; s/ .*\$//')
+    ${task.process}:
+        python: \$(python --version | sed 's/Python //g')
     END_VERSIONS
     """
 }
